feat: add table order ticket and fix price event

+++ MyTicket.WebApi/Endpoints/Event/Models/Response/GetEventResponse.cs
@@ -8,9 +8,9 @@ public class GetEventResponse
     public DateTime StartDate { get; set; }
     public DateTime EndDate { get; set; }
     public int AvailableTickets { get; set; }
-    public int Price { get; set; }
+    public decimal Price { get; set; }
     public string Location { get; set; }
-    
+
     public string? CreatedBy { get; set; }
     public string? CreatedByName { get; set; }
     public string? CreatedByFullName { get; set; }
